+++ cluster_agent/jobbergate/schemas.py
@@ -1,0 +1,14 @@
+import pydantic
+
+
+class PendingJobSubmission(pydantic.BaseModel, extra=pydantic.Extra.ignore):
+    """
+    Specialized model for the cluster-agent to pull a pending job_submission along with
+    data from its job_script and application sources.
+    """
+
+    id: int
+    job_submission_name: str
+    job_script_name: str
+    job_script_data_as_string: str
+    application_name: str